refact: add usePaginate hooks to handle pagination logic
